requirements.txt
langchain==0.3.27
langchain-anthropic==0.3.18
langchain-community==0.3.27
langchain-core==0.3.74
langchain-groq==0.3.7
langchain-openai==0.3.29
langchain-text-splitters==0.3.9
langchain-together==0.3.1
langsmith==0.4.13
openai==1.99.9
openpyxl==3.1.5
pymilvus==2.5.14
PyPDF2==3.0.1
python-docx==1.2.0
sentence-transformers==5.1.0
streamlit==1.31.1
langchain-mistralai==0.2.11
google-genai==1.30.0
pdf2image==1.16.3
pillow==10.0.0